requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.12=h4ab18f5_0
aom=3.9.1=hac33072_0
blas=1.0=openblas
blosc=1.21.6=hef167b5_0
brotli=1.0.9=he6710b0_2
brunsli=0.1=h2531618_0
bzip2=1.0.8=h5eee18b_6
c-ares=1.34.2=heb4867d_0
c-blosc2=2.15.1=hc57e6cf_0
ca-certificates=2024.9.24=h06a4308_0
cairo=1.18.0=hebfffa5_3
charls=2.4.2=h59595ed_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
cyrus-sasl=2.1.27=h54b06d7_7
dav1d=1.2.1=hd590300_0
dbus=1.13.6=h5008d03_3
double-conversion=3.3.0=h59595ed_0
expat=2.6.3=h5888daf_0
ffmpeg=7.1.0=gpl_he2fd91e_702
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.54.1=pypi_0
freeglut=3.2.2=ha6d2627_3
freetype=2.12.1=h267a509_2
fribidi=1.0.10=h36c2ea0_0
gdk-pixbuf=2.42.12=hb9ae30d_0
giflib=5.2.2=h5eee18b_0
gmp=6.3.0=hac33072_2
graphite2=1.3.13=h59595ed_1003
harfbuzz=9.0.0=hda332d3_1
hdf5=1.14.4=nompi_h2d575fe_101
icu=75.1=he02047a_0
imagecodecs=2024.9.22=py310hdbe5077_0
imageio=2.33.1=py310h06a4308_0
imath=3.1.12=h7955e40_0
jasper=4.2.4=h536e39c_0
jxrlib=1.1=h7b6447c_2
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=pypi_0
krb5=1.21.3=h659f571_0
lame=3.100=h166bdaf_1003
lazy_loader=0.4=py310h06a4308_0
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h27087fc_0
libabseil=20240722.0=cxx17_h5888daf_1
libaec=1.1.3=h59595ed_0
libasprintf=0.22.5=he8f35ee_3
libass=0.17.3=h1dc1e6a_0
libavif16=1.1.1=h104a339_1
libblas=3.9.0=25_linux64_openblas
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=25_linux64_openblas
libclang-cpp19.1=19.1.2=default_hb5137d0_1
libclang13=19.1.2=default_h9c6a7e4_1
libcups=2.3.3=h4637d8d_4
libcurl=8.10.1=hbbe4b11_0
libdeflate=1.22=hb9d3cd8_0
libdrm=2.4.123=hb9d3cd8_0
libedit=3.1.20191231=he28a2e2_2
libegl=1.7.0=ha4b6fd6_1
libev=4.33=hd590300_2
libexpat=2.6.3=h5888daf_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgettextpo=0.22.5=he02047a_3
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgl=1.7.0=ha4b6fd6_1
libglib=2.82.2=h2ff4ddf_0
libglu=9.0.0=ha6d2627_1004
libglvnd=1.7.0=ha4b6fd6_1
libglx=1.7.0=ha4b6fd6_1
libgomp=14.2.0=h77fa898_1
libhwloc=2.11.1=default_hecaa2ac_1000
libhwy=1.1.0=h00ab1b0_0
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.0=hd590300_1
libjxl=0.11.0=hdb8da77_2
liblapack=3.9.0=25_linux64_openblas
liblapacke=3.9.0=25_linux64_openblas
libllvm19=19.1.2=ha7bfdaf_0
libnghttp2=1.64.0=h161d5f1_0
libnsl=2.0.1=hd590300_0
libntlm=1.4=h7f98852_1002
libopenblas=0.3.28=pthreads_h94d23a6_0
libopencv=4.10.0=qt6_py310hf9d9ec8_610
libopenvino=2024.4.0=hac27bb2_2
libopenvino-auto-batch-plugin=2024.4.0=h4d9b6c2_2
libopenvino-auto-plugin=2024.4.0=h4d9b6c2_2
libopenvino-hetero-plugin=2024.4.0=h3f63f65_2
libopenvino-intel-cpu-plugin=2024.4.0=hac27bb2_2
libopenvino-intel-gpu-plugin=2024.4.0=hac27bb2_2
libopenvino-intel-npu-plugin=2024.4.0=hac27bb2_2
libopenvino-ir-frontend=2024.4.0=h3f63f65_2
libopenvino-onnx-frontend=2024.4.0=h5c8f2c3_2
libopenvino-paddle-frontend=2024.4.0=h5c8f2c3_2
libopenvino-pytorch-frontend=2024.4.0=h5888daf_2
libopenvino-tensorflow-frontend=2024.4.0=h6481b9d_2
libopenvino-tensorflow-lite-frontend=2024.4.0=h5888daf_2
libopus=1.3.1=h7f98852_1
libpciaccess=0.18=hd590300_0
libpng=1.6.44=hadc24fc_0
libpq=17.0=h04577a9_4
libprotobuf=5.28.2=h5b01275_0
librsvg=2.58.4=hc0ffecb_0
libsqlite=3.47.0=hadc24fc_0
libssh2=1.11.0=h0841786_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libtiff=4.7.0=he137b08_1
libuuid=2.38.1=h0b41bf4_0
libva=2.22.0=h8a09558_1
libvpx=1.14.1=hac33072_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h2c5496b_1
libxml2=2.12.7=he7c6b58_4
libzlib=1.3.1=hb9d3cd8_2
libzopfli=1.0.3=he6710b0_0
lz4-c=1.9.4=h6a678d5_1
matplotlib=3.9.2=pypi_0
mysql-common=9.0.1=h266115a_2
mysql-libs=9.0.1=he0572af_2
ncurses=6.4=h6a678d5_0
networkx=3.3=py310h06a4308_0
nptyping=2.5.0=pypi_0
numpy=1.26.4=py310heeff2f4_0
numpy-base=1.26.4=py310h8a23956_0
ocl-icd=2.3.2=hd590300_1
opencv=4.10.0=qt6_py310h6786a00_610
openexr=3.3.1=hccdc605_2
openh264=2.4.1=h59595ed_0
openjpeg=2.5.2=he7f1fd0_0
openldap=2.6.8=hedd0468_0
openssl=3.3.2=hb9d3cd8_0
packaging=24.1=py310h06a4308_0
pango=1.54.0=h4c5309f_1
pcre2=10.44=hba22ea6_2
pillow=11.0.0=py310hfeaa1f3_0
pip=24.2=py310h06a4308_0
pixman=0.43.2=h59595ed_0
pthread-stubs=0.4=hb9d3cd8_1002
pugixml=1.14=h59595ed_0
py-opencv=4.10.0=qt6_py310h3e0c198_610
pybind11-abi=4=hd3eb1b0_1
pynrrd=1.0.0=pypi_0
pyparsing=3.2.0=pypi_0
python=3.10.13=hd12c33a_1_cpython
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.10=5_cp310
qt6-main=6.7.3=h6e8976b_1
rav1e=0.6.6=he8a937b_2
readline=8.2=h5eee18b_0
scienceplots=2.1.1=pypi_0
scikit-image=0.24.0=py310h1128e8f_0
scipy=1.13.1=py310heeff2f4_0
setuptools=75.1.0=py310h06a4308_0
six=1.16.0=pypi_0
snappy=1.2.1=ha2e4443_0
sqlite=3.32.3=hcee41ef_1
svt-av1=2.2.1=h5888daf_0
tbb=2021.13.0=h84d6215_0
tifffile=2023.4.12=py310h06a4308_0
tk=8.6.13=noxft_h4845f30_101
typing-extensions=4.12.2=pypi_0
tzdata=2024b=h04d1e81_0
wayland=1.23.1=h3e06ad9_0
wayland-protocols=1.37=hd8ed1ab_0
wheel=0.44.0=py310h06a4308_0
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xcb-util=0.4.1=hb711507_2
xcb-util-cursor=0.1.5=hb9d3cd8_0
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.43=hb9d3cd8_0
xorg-libice=1.1.1=hb9d3cd8_1
xorg-libsm=1.2.4=he73a12e_1
xorg-libx11=1.8.10=h4f16b4b_0
xorg-libxau=1.0.11=hb9d3cd8_1
xorg-libxcomposite=0.4.6=hb9d3cd8_2
xorg-libxcursor=1.2.2=hb9d3cd8_0
xorg-libxdamage=1.1.6=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.1=hb9d3cd8_0
xorg-libxi=1.8.2=hb9d3cd8_0
xorg-libxrandr=1.5.4=hb9d3cd8_0
xorg-libxrender=0.9.11=hb9d3cd8_1
xorg-libxtst=1.2.5=hb9d3cd8_3
xorg-libxxf86vm=1.1.5=hb9d3cd8_4
xorg-xextproto=7.3.0=hb9d3cd8_1004
xorg-xorgproto=2024.1=hb9d3cd8_1
xz=5.4.6=h5eee18b_1
zfp=1.0.1=h5888daf_2
zlib=1.3.1=hb9d3cd8_2
zlib-ng=2.2.2=h5888daf_0
zstd=1.5.6=ha6fb4c9_0
